backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.4
reportlab==4.0.9
Pillow==10.1.0
gunicorn==21.2.0
openpyxl==3.1.2
